Refactor sagemaker_notebook into sagemaker files
